Add APIConfig Struct (#112)
This merge adds an `APIConfig` struct to collect the parameters required by an `APIClient`. The `NewAPIClient` function has been updated to use this new struct, and any callers of `NewAPIClient` have been updated to use an `APIConfig` structure.

`NewAPIClient` tests were updated to use this structure, but validation was not updated, so that we could verify that behavior remained consistent.

As a bit of cleanup, TLS config and setup were moved in the api client file, so that API concerns and TLS/HTTP Transport concerns were closer to similar code. Previously, they were interspersed.