Refactor Version Code for Consistency with Other Products (#124)
This merge updates the version code to be more consistent with other HashiCorp products. It is an internal change and should not have user impact. However, it is one step in the process that would allow us to release more frequently, and it also sets the stage to later include detailed version information in Manifest.json output files.