Assign Unique Names to Temporary Files & Store them in Destination Directory (#134)
This merge enables concurrent runs of hcdiag, such as in functional tests, by ensuring temporary directories are uniquely named. The temporary directories will have a unique string appended to their names, ensuring that even if multiple hcdiag runs occur at the same time, a separate directory will be created for each one.

In addition, we also now use the destination directory as our working directory, rather than the current directory. By default, the behavior will be the same - because the current directory is also the destination, our temporary directory will be in the current working directory. However, when this value is overridden, the temporary directory will be placed within the destination directory instead of the current directory.