Print Summary to Standard Out at end of Agent Run (#141)
This merge adds a summary report to the end of an hcdiag agent run, with the intention of making it a bit easier for users to quickly see the overall status of seekers. The report is a simple table, which includes - for each product that was part of the run - a count of successful, failed, and unknown seekers, as well as the total number of seekers.

Unlike log messages, the report goes to Standard Out, meaning the run can be piped to an output file if desired.