Display Summary Report Alphabetically by Product (#143)
Report output is now sorted alphabetically by product name, rather than nondeterministically. This should be a better user experience, knowing where a product should appear in the report. And, it should fix sporadic unit test failures, which had been written with an (incorrect) assumption that the order of reports would be deterministic already.