Initial Support for Product Auto-Detection (#152)
This merge adds support for product auto-detection. The hope is to make it simpler for users to get diagnostics on all installed products by simply running the `hcdiag` command. This is similar to the previous `all` flag, but with some basic checks in place to validate which products' CLIs are actually installed first.