Allow Redactions on Host Runners (#192)
This merge enables redactions on the host runners that leverage third-party Go modules for data collection. Specifically, redaction is now supported for the following runners:

* host.Disk
* host.Info
* host.Network
* host.Process

Secondarily, it also enables redact.String(), redact.JSON(), and redact.Bytes() to more quickly return the input object, unchanged, when the list of redactions is empty.