clean up temporary directory regardless of whether hcdiag succeeds or fails (#235)
This addresses a bug, in which the temporary directory created by the agent would only be cleaned up in the event that hcdiag ran to its expected completion. Certain error cases, such as product healthcheck failures, would end up leaving the temporary directory on the filesystem. We now should clean up after ourselves, regardless of whether the agent run completes successfully or whether it errors.