Improve Handling of Commander Args with Spaces (#239)
This intends to improve hcdiag's ability to handle commander args that include spaces in them on Linux/Unix OS's. Previously, such arguments would be inappropriately split, even if they were wrapped within single quotes, which caused the command to be unable to make sense of the arguments. A library has been included which should alleviate this issue, and tests were added around the capability.

Note that no changes are made to Windows command argument processing. The underlying mechanism for Windows command calls is done differently than *nix, so the existing behavior worked for Windows.