Add Bundle Location to End-of-Run Summary (#241)
This merge adds a message indicating the run completed, along with a notification of where the bundle was written, to the end of the diagnostic run. The hope is this will make it easier for users to locate the bundle.

Along with this, the merge moves the responsibility of displaying the summary from the agent to the command package. This was intended to maintain the CLI as the intermediary between the user and the agent.