Enable Context Propagation from Agent to Products (#263)
This merge adds a context field to Agent, which is then propagated downward to products. This sets the stage for implementation of context and timeouts/cancelation within runners. It also allows us to set custom timeouts/context within the CLI as well, by enabling custom context to be set on Agent.

Wherever exported methods/functions were updated to support context, the merge does this by migrating most logic to a "WithContext" function and then simply wrapping it with a background context in the previously existing function. This was done intentionally to maintain backwards compatibility.

Where unexported functions were updated, the function signatures were simply updated to now include a context.