Add Timeout & Cancelation to Command Runner (#270)
This merge enables timeouts and cancellation in the Command runner. Timeouts may be passed in via optional configuration. If the timeout expires, or if a context that has been passed into the Command runner when it was constructed is canceled, then the run will stop with an appropriate operation status type.