Migrate HTTP Runner Constructors to Use Config Objects (#273)
This merge updates the HTTP constructors to leverage a new HttpConfig object, rather than parameters. The intention is to simplify future feature additions by limiting changes to the function signatures of these constructors.

Unrelated to the primary change, a configuration setting required updating for integration tests to pass using a new version of Consul that was recently released, version 1.14.0. This change to Consul does not impact hcdiag, but prevented the dev Consul server from starting during integration tests.