Add Timeout & Cancellation to Journald Log Runner (#284)
This merge enables timeouts and cancellation in the Journald Log runner. Timeouts may be passed in via optional configuration. If the timeout expires, or if a context that has been passed into the runner when it was constructed is canceled, then the run will stop with an appropriate operation status type.