Warn for Missing Vault Tokens Instead of Failing (#294)
This merge allows hcdiag to execute a diagnostic run for Vault even when the Vault token cannot be found. Previously, this pre-flight check would stop the entire hcdiag execute, however some useful information could still be gleaned even without the token. Now, we write a warning (an error-level log message, which should make it red in terminals that support color) to inform users that the data will be incomplete and to suggest how they might address this. But, we still create a bundle, so that we can collect some information to provide to Support.